(1814) modification schemas et labels
