+++ packages/frontend/mobile/src/js/components/FormMiseAJourDeSiteSectionConditionsDeVie/inputs/InputWaterAccessComments.vue
@@ -12,5 +12,5 @@
 
 <script setup>
 import { TextArea } from "@resorptionbidonvilles/ui";
-import labels from "../FormMiseAJourDeSiteSectionConditionsDeVie.labels";
+import labels from "#src/js/pages/MiseAjourSectionDeSite/MiseAJourSectionDeSite.labels";
 </script>
