--- packages/frontend/mobile/src/js/components/FormMiseAjourDeSiteSectionHabitants/FormMiseAJourDeSiteSectionHabitants.labels.js
@@ -1,17 +1,0 @@
-export default {
-    population_total: "Personnes",
-    population_couples: "Ménages",
-    population_minors: "Mineurs",
-    population_minors_0_3: "0 - 3 ans",
-    population_minors_3_6: "3 - 6 ans",
-    population_minors_6_12: "6 - 12 ans",
-    population_minors_12_16: "12 - 16 ans",
-    population_minors_16_18: "16 - 18 ans",
-    minors_in_school: "Inscrits en établissement scolaire",
-    social_origins: "Quelle est l'origine des habitants ?",
-    caravans: "Caravanes",
-    huts: "Habitats autoconstruits",
-    tents: "Tentes",
-    cars: "Voitures dortoir",
-    mattresses: "Matelas",
-};
